修复AHCI的bug

+++ user/ahcid/sata.h
@@ -5,7 +5,7 @@
 #define SATA_REG_FIS_D2H 0x34
 #define SATA_REG_FIS_H2D 0x27
 #define SATA_REG_FIS_COMMAND 0x80
-#define SATA_LBA_COMPONENT(lba, offset) ((((lba) >> (offset)) & 0xff))
+#define SATA_LBA_COMPONENT(lba, offset) ((uint8_t)(((lba) >> (offset)) & 0xff))
 
 #define ATA_IDENTIFY_DEVICE 0xec
 #define ATA_IDENTIFY_PAKCET_DEVICE 0xa1
